% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-withrcpp.Rmd.
\name{alternate_signs}
\alias{alternate_signs}
\title{Alternately flip the signs of elements of a vector}
\usage{
alternate_signs(v)
}
\arguments{
\item{v}{A numerical vector}
}
\description{
Alternately flip the signs of elements of a vector
}