% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-withrcpp.Rmd.
\docType{package}
\name{withrcpp-package}
\alias{withrcpp}
\alias{withrcpp-package}
\title{My great package}
\description{
Here's an overview of the package.
}
\seealso{
\code{\link{alternate_signs}}
}
\author{
\strong{Maintainer}: First Last \email{you@gmail.com}

Authors:
\itemize{
  \item First Last \email{you@gmail.com}
}

}
\keyword{internal}